Pre-allocate memory and mutate it directly
This saves a ton of intermediate Vec<u8> objects from being created.
The optimization of using Rayon did give parallelization, but it also
created a ton of overhead in having to gather up small Vecs of u8s to
describe one pixel at a time.

The previous PR that nested two parallel iterators inside of each other
also had a similar savings, because when an entire row of pixels was
created, they could easily throw out all of the temporary Vec objects
that held them. However by moving back to using one giant parallel
iterator, we cannot deallocate those objects until the entire program is
complete. By avoiding the allocaations in the first place, we save a lot
of memory.